+++ packages/documentation-framework/scripts/cli/build.js
@@ -35,7 +35,7 @@ async function buildWebpack(webpackConfig) {
       }
     
       if (stats.hasWarnings()) {
-        console.warn(info.warnings.join('\n'));
+        console.warn(info.warnings.map(w => w.message || w).join('\n'));
       }
 
       res();
@@ -76,7 +76,7 @@ async function execFile(file, args) {
 }
 
 async function build(cmd, options) {
-  generate(options);
+  await generate(options);
   const toBuild = cmd === 'all'
     ? ['server', 'client']
     : cmd;