+++ packages/documentation-framework/scripts/cli/generate.js
@@ -1,15 +1,17 @@
 const path = require('path');
-const { sourceMD, sourceProps, sourceFunctionDocs, writeIndex } = require('../md/parseMD');
+const { sourceMD, sourceProps, sourceFunctionDocs, writeIndex, waitForProps, processMD } = require('../md/parseMD');
 
 function getSource(options) {
   return require(path.join(process.cwd(), options.parent.source));
 }
 
-function generate(options) {
+async function generate(options) {
   const start = new Date();
   console.log('write source files to patternfly-docs/generated');
   const sourceMDWithOptions = (glob, source, ignore) => sourceMD(glob, source, ignore, options._name);
   getSource(options)(sourceMDWithOptions, sourceProps, sourceFunctionDocs);
+  await waitForProps();
+  processMD();
   const exitCode = writeIndex();
   if (exitCode !== 0) {
     process.exit(exitCode);