+++ packages/documentation-framework/scripts/tsDocgen.js
@@ -1,6 +1,14 @@
 const fs = require("fs");
-const reactDocgen = require("react-docgen");
 const ts = require("typescript");
+
+// react-docgen v6+ is ESM-only; lazy-load it asynchronously on first use
+let reactDocgenReady = null;
+function loadReactDocgen() {
+  if (!reactDocgenReady) {
+    reactDocgenReady = import("react-docgen");
+  }
+  return reactDocgenReady;
+}
 
 const annotations = [
   {
@@ -41,15 +49,14 @@ function addAnnotations(prop) {
   return prop;
 }
 
-function getComponentMetadata(filename, sourceText) {
+async function getComponentMetadata(filename, sourceText) {
+  const reactDocgen = await loadReactDocgen();
   let parsedComponents = null;
   try {
-    parsedComponents = reactDocgen.parse(
-      sourceText,
-      reactDocgen.resolver.findAllExportedComponentDefinitions,
-      null,
-      { filename }
-    );
+    parsedComponents = reactDocgen.parse(sourceText, {
+      resolver: new reactDocgen.builtinResolvers.FindExportedDefinitionsResolver(),
+      filename
+    });
   } catch (err) {
     // eslint-disable-next-line no-console
     // console.warn(`No component found in ${filename}:`, err);
@@ -161,9 +168,9 @@ function normalizeProp([
   return res;
 }
 
-function tsDocgen(file) {
+async function tsDocgen(file) {
   const sourceText = fs.readFileSync(file, "utf8");
-  const componentMeta = getComponentMetadata(file, sourceText); // Array of components with props
+  const componentMeta = await getComponentMetadata(file, sourceText); // Array of components with props
   const interfaceMeta = getInterfaceMetadata(file, sourceText); // Array of interfaces with props
   const typeAliasMeta = getTypeAliasMetadata(file, sourceText); // Array of type aliases with props
   const propsMetaMap = [...interfaceMeta, ...typeAliasMeta].reduce(function (